requirements.txt
psycopg2==2.9.9
streamlit==1.38.0
sentence_transformers==2.7.0
openai==1.42.0
python-dotenv==1.0.1
elasticsearch==8.14.0
pandas==2.2.2
